requirements.txt
asttokens==2.4.1
boto3==1.33.13
botocore==1.33.13
certifi==2024.12.14
channelnorm-cuda==0.0.0
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
correlation-cuda==0.0.0
cycler==0.11.0
deep-point==1.0
docopt==0.6.2
environs==10.1.0
executing==2.0.1
fonttools==4.38.0
humanfriendly==10.0
humanize==4.6.0
icecream==2.1.3
idna==3.10
importlib-metadata==6.7.0
jmespath==1.0.1
kiwisolver==1.4.5
llvmlite==0.39.1
marshmallow==3.19.0
matplotlib==3.5.3
multiprocessing-logging==0.3.4
ninja==1.11.1.1
numba==0.56.4
numpy==1.21.6
opencv-python==4.9.0.80
packaging==24.0
petrel-oss-sdk===v2.2.1-2-g1505ef3-master
Pillow==9.5.0
pipreqs==0.4.13
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
resample2d-cuda==0.0.0
s3transfer==0.8.2
scipy==1.3.1
six==1.16.0
torch==1.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.66.2
typing_extensions==4.7.1
urllib3==1.26.20
yarg==0.1.10
zipp==3.15.0
